Dockmaster

Description

Runtime execution

dockmaster

Arguments

:: FilePath
Composition

-> Bool
Local-only flag

-> Text
Command

-> [Text]
Opts/args for Command

-> Sh ()

Runs docker-compose commands against resolved composition locations See usage docs for more info. Tries to find a dockmaster.yml file based on the initial path argument

dockmaster.yml

data Dockmaster

Dockmaster configuration (specified by dockmaster.yml)

Constructors

Dockmaster

Fields

dmCompose :: ComposeConfig

dmEnv :: EnvConfig

dmTargets :: [Target]

dmCommands :: HashMap Text CommandConfig

data ComposeConfig

Configuration for docker-compose.yml arguments.

Constructors

ComposeConfig

Fields

dcFiles :: [ComposeFile]

dcFlags :: [Text]

data ComposeFile

Configuration for docker-compose compose file construction.

Constructors

ComposeFile

Fields

cfPath :: FilePath

cfTemplate :: Bool

cfConfig :: [FilePath]

data EnvConfig

Global environment variables for dockmaster execution.

Constructors

EnvConfig

Fields

ecFiles :: [FilePath]

ecVars :: HashMap Text Text

data Target

Targets are used to identify where compositions are run.

Constructors

Target

Fields

targetName :: Text

targetType :: Text

targetMachine :: Maybe Text

data Hook

Hooks can be specified by filename or direct shell command.

Constructors

File Text

Shell Text

data CommandConfig

Configuration for each command.

Constructors

CommandConfig

Fields

ccRunCompose :: Bool

ccPreHooks :: [Hook]

ccPostHooks :: [Hook]

isTemplate :: Maybe String -> Bool

Small helper method to parse dockmastercomposefiles/type.

Dockmaster yml parsing

dockmasterYml :: Sh (Either Text Dockmaster)

Parse $CWD/dockmaster.yml

Note this assumes we are already in the correct dockmaster workdir.

parseYml :: FilePath -> Sh (Either Text Dockmaster)

Parse dockmaster yml file

Dockmaster interpretation

prepareEnv :: Dockmaster -> Sh ()

Sets environment variables defined by dockmaster configuration

dockermachine :: Text -> Sh a -> Sh a

Takes machine name and Sh action, and wraps Sh action in scope of docker-machine env

hookWrap :: Text -> Sh () -> Sh ()

Executes specific dc command pre/post hooks around action argument (action arg is typically docker-compose command).

hookWrap' :: Dockmaster -> Text -> Sh () -> Sh ()

Same thing as hookWrap but accepts Dockmaster config to execute against.

execHook :: Hook -> Sh ()

Execute hook

Orchestration for docker-compose.yml

dockercompose :: Dockmaster -> [Text] -> Sh ()

Runs docker-compose but uses Dockmaster configuration to possibly

Provide multiple templated docker-compose.yml files

Pass global docker-compose flags on every call

config.yml

data Config

Dockmaster configuration

Constructors

Config

Fields

dmcPaths :: [FilePath]

Error types

data DmcError

Dockmaster configuration runtime errors

Constructors

WorkDirNotFound

DecodingError Text

Getting global config

config :: Sh (Either DmcError Config)

Get global dockmaster config

If config.yml fails to parse, returns a Left error. If config.yml is not found, returns Right baseConfig (default configuration).

baseConfig :: Config

Get base config options

Resolving relative paths

getWorkDir :: FilePath -> Sh (Either DmcError FilePath)

Resolve the appropriate dockmaster workdir.

For example, if $CWD/dockmaster.yml exists, then

>>> getWorkDir "."
Right "."

If $CWD/dockmaster.yml does not exist, then

>>> getWorkDir "."
Left "dockmaster.yml file not found"

This function will also try to resolve relative paths against the dmcPaths composition listing directories, if any are specified by global config. For example, if:

$HOME/git is a PATH specified in global config.yml

$HOME/git/deploybot/dockmaster.yml exists

$CWD/deploybot/dockmaster.yml does not exist, then

>>> getWorkDir "deploybot"
Right "$HOME/git/deploybot"

getWorkDir' :: Config -> FilePath -> Sh (Either DmcError FilePath)

Same thing as getWorkDir but uses a Config argument instead of resolving one.