Dockmaster.Config.Parser

Description

Getting global config

config :: Sh (Either Text Config)

Get global dockmaster config

If config.yml fails to parse, returns a Left error. If config.yml is not found, returns Right baseConfig (default configuration).

baseConfig :: Config

Get base config options

Resolving relative paths

getWorkDir :: FilePath -> Sh (Either Text FilePath)

Resolve the appropriate dockmaster workdir.

For example, if $CWD/dockmaster.yml exists, then

>>> getWorkDir "."
Right "."

If $CWD/dockmaster.yml does not exist, then

>>> getWorkDir "."
Left "dockmaster.yml file not found"

This function will also try to resolve relative paths against the dmcPath composition listing directories, if any are specified by global config. For example, if:

$HOME/git is a PATH specified in global config.yml

$HOME/git/deploybot/dockmaster.yml exists

$CWD/deploybot/dockmaster.yml does not exist, then

>>> getWorkDir "deploybot"
Right "$HOME/git/deploybot"

getWorkDir' :: Config -> FilePath -> Sh (Either Text FilePath)

Same thing as getWorkDir but uses a Config argument instead of resolving one.

Re-exported for convenience

data Config

Dockmaster configuration

Constructors

Config

Fields

dmcPaths :: [FilePath]