Dockmaster.Parser

Description

Documentation

dockmasterYml :: Sh (Either Text Dockmaster)

Parse $CWD/dockmaster.yml

Note this assumes we are already in the correct dockmaster workdir.

Dockmaster command configuration

hookWrap :: Text -> Sh () -> Sh ()

Executes specific dc command pre/post hooks around action argument (action arg is typically docker-compose command).

hookWrap' :: Dockmaster -> Text -> Sh () -> Sh ()

Same thing as hookWrap but accepts Dockmaster config to execute against.

Re-exported for convenience

data Dockmaster

Dockmaster configuration (specified by dockmaster.yml)

Constructors

Dockmaster

Fields

dmFile :: Maybe ComposeFile

dmTargets :: [Target]

dmCommands :: HashMap Text CommandConfig

data ComposeFile

Configuration for docker-compose.yml file template & vars

Constructors

ComposeFile

Fields

cfPath :: FilePath

cfVars :: [FilePath]

data Target

Targets are used to identify where compositions are run

Constructors

Target

Fields

targetName :: Text

targetType :: Text

targetMachine :: Maybe Text

data Hook

Hooks can be specified by filename or direct shell command

Constructors

File Text

Shell Text

data CommandConfig

Configuration for each command

Constructors

CommandConfig

Fields

ccCompose :: Bool

ccPreHooks :: [Hook]

ccPostHooks :: [Hook]