Dockmaster

Description

Runtime execution

dm :: FilePath -> Text -> [Text] -> Sh ()

Runs docker-compose commands against resolved composition locations See usage docs for more info. Tries to find a dockmaster.yml file based on the initial path argument

dockmaster.yml

data Dockmaster

Dockmaster configuration (specified by dockmaster.yml)

Constructors

Dockmaster

Fields

dmFile :: Maybe ComposeFile

dmTargets :: [Target]

dmCommands :: HashMap Text CommandConfig

data ComposeFile

Configuration for docker-compose.yml file template & vars.

Constructors

ComposeFile

Fields

cfPath :: FilePath

cfVars :: [FilePath]

data Target

Targets are used to identify where compositions are run.

Constructors

Target

Fields

targetName :: Text

targetType :: Text

targetMachine :: Maybe Text

data Hook

Hooks can be specified by filename or direct shell command.

Constructors

File Text

Shell Text

data CommandConfig

Configuration for each command.

Constructors

CommandConfig

Fields

ccCompose :: Bool

ccPreHooks :: [Hook]

ccPostHooks :: [Hook]

dockmasterYml :: Sh (Either Text Dockmaster)

Parse $CWD/dockmaster.yml

Note this assumes we are already in the correct dockmaster workdir.

Dockmaster command configuration

hookWrap :: Text -> Sh () -> Sh ()

Executes specific dc command pre/post hooks around action argument (action arg is typically docker-compose command).

hookWrap' :: Dockmaster -> Text -> Sh () -> Sh ()

Same thing as hookWrap but accepts Dockmaster config to execute against.

Orchestration for docker-compose.yml

dockercompose :: Dockmaster -> [Text] -> Sh ()

Runs docker-compose but uses compiled docker-compose.yml file if template/vars are specified in dockmaster.yml.

config.yml

data Config

Dockmaster configuration

Constructors

Config

Fields

dmcPaths :: [FilePath]

Getting global config

config :: Sh (Either Text Config)

Get global dockmaster config

If config.yml fails to parse, returns a Left error. If config.yml is not found, returns Right baseConfig (default configuration).

baseConfig :: Config

Get base config options

Resolving relative paths

getWorkDir :: FilePath -> Sh (Either Text FilePath)

Resolve the appropriate dockmaster workdir.

For example, if $CWD/dockmaster.yml exists, then

>>> getWorkDir "."
Right "."

If $CWD/dockmaster.yml does not exist, then

>>> getWorkDir "."
Left "dockmaster.yml file not found"

This function will also try to resolve relative paths against the dmcPath composition listing directories, if any are specified by global config. For example, if:

$HOME/git is a PATH specified in global config.yml

$HOME/git/deploybot/dockmaster.yml exists

$CWD/deploybot/dockmaster.yml does not exist, then

>>> getWorkDir "deploybot"
Right "$HOME/git/deploybot"

getWorkDir' :: Config -> FilePath -> Sh (Either Text FilePath)

Same thing as getWorkDir but uses a Config argument instead of resolving one.